User experience improvements
Use the prompt toolkit instead of `raw_input()` to provide a better
interactive mode experience. Leverage the recently implemented
SignalFlow Pygments lexer to provide syntax highlighting of programs.

Also support executing from a program file with the `-x` option, which
allows for writing program files with a `#!signalflow` shebang.

Look for a session token in ~/.sftoken if no token is provided before
asking the user to login.